+++ SampleCode/PingOneVerify/app/src/main/java/com/pingidentity/sdk/pingoneverify/ui/fragments/DocumentCaptureDialog.java
@@ -17,7 +17,7 @@ import com.pingidentity.sdk.pingoneverify.sample.R;
 import com.pingidentity.sdk.pingoneverify.sample.databinding.DialogDocumentCaptureBinding;
 import com.pingidentity.sdk.pingoneverify.ui.UiConstants;
 import com.pingidentity.sdk.pingoneverify.ui.utils.BindingAdapters;
-import com.pingidentity.sdk.pingoneverify.utils.VerifySessionTimer;
+import com.pingidentity.sdk.pingoneverify.utils.DocumentSubmissionTimer;
 import com.pingidentity.sdk.provider.language.LanguagePackProviderContract;
 
 public class DocumentCaptureDialog extends BaseFragment {
@@ -45,8 +45,8 @@ public class DocumentCaptureDialog extends BaseFragment {
     }
 
     @Nullable
-    private final VerifySessionTimer.Timer.TimerObserver mTimerObserver =
-            new VerifySessionTimer.Timer.TimerObserver() {
+    private final DocumentSubmissionTimer.Observer mTimerObserver =
+            new DocumentSubmissionTimer.Observer() {
                 @Override public void onTick(int millisRemaining) {
                     if (mBinding != null) mBinding.txtTimeRemaining.setText(getTimeRemainingString(millisRemaining, R.string.idv_timer_label));
                 }
@@ -68,7 +68,7 @@ public class DocumentCaptureDialog extends BaseFragment {
         setStartButton(mBinding);
         setScreenContent(mBinding);
 
-        VerifySessionTimer.getInstance().getSessionTimer().addObserver(mTimerObserver);
+        DocumentSubmissionTimer.getInstance().addObserver(mTimerObserver);
 
         mBinding.btnCancel.setOnClickListener(this::onCancel);
         setDocumentClass(mDocument);
@@ -104,7 +104,7 @@ public class DocumentCaptureDialog extends BaseFragment {
     @Override
     public void onDestroyView() {
         super.onDestroyView();
-        VerifySessionTimer.getInstance().getSessionTimer().removeObserver(mTimerObserver);
+        DocumentSubmissionTimer.getInstance().removeObserver(mTimerObserver);
         mBinding = null;
     }