+++ SampleCode/PingOneVerify/app/src/main/java/com/pingidentity/sdk/pingoneverify/ui/fragments/DocumentSelectFragment.kt
@@ -11,7 +11,7 @@ import com.pingidentity.sdk.pingoneverify.sample.R
 import com.pingidentity.sdk.pingoneverify.sample.databinding.DialogDocumentSelectBinding
 import com.pingidentity.sdk.pingoneverify.ui.views.DocumentAdapter
 import com.pingidentity.sdk.pingoneverify.ui.views.DocumentCallback
-import com.pingidentity.sdk.pingoneverify.utils.VerifySessionTimer
+import com.pingidentity.sdk.pingoneverify.utils.DocumentSubmissionTimer
 import com.pingidentity.sdk.provider.language.LanguagePackProviderContract
 
 class DocumentSelectFragment(
@@ -22,7 +22,7 @@ class DocumentSelectFragment(
     private lateinit var binding: DialogDocumentSelectBinding
 
 
-    private val mTimerObserver = object : VerifySessionTimer.Timer.TimerObserver {
+    private val mTimerObserver = object : DocumentSubmissionTimer.Observer {
         override fun onTick(millisRemaining: Int) {
             binding?.txtTimeRemaining?.text = getTimeRemainingString(millisRemaining, R.string.idv_timer_label)
         }
@@ -31,7 +31,7 @@ class DocumentSelectFragment(
 
     override fun onDestroyView() {
         super.onDestroyView()
-        VerifySessionTimer.getInstance().getSessionTimer().removeObserver(mTimerObserver)
+        DocumentSubmissionTimer.getInstance().removeObserver(mTimerObserver)
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
@@ -50,7 +50,7 @@ class DocumentSelectFragment(
                 layoutManager = LinearLayoutManager(context)
                 adapter = documentsAdapter
             }
-            VerifySessionTimer.getInstance().getSessionTimer().addObserver(mTimerObserver)
+            DocumentSubmissionTimer.getInstance().addObserver(mTimerObserver)
             binding.btnCancel.setOnClickListener { onBackPressed() }
         }
     }