+++ SampleCode/PingOneVerify/app/src/main/java/com/pingidentity/sdk/pingoneverify/ui/fragments/LocationFragment.kt
@@ -11,7 +11,7 @@ import com.pingidentity.sdk.pingoneverify.neo.models.DocumentClass
 import com.pingidentity.sdk.pingoneverify.sample.R
 import com.pingidentity.sdk.pingoneverify.sample.databinding.DialogLocationBinding
 import com.pingidentity.sdk.pingoneverify.neo.settings.LocationCaptureSettings
-import com.pingidentity.sdk.pingoneverify.utils.VerifySessionTimer
+import com.pingidentity.sdk.pingoneverify.utils.DocumentSubmissionTimer
 import com.pingidentity.sdk.provider.language.LanguagePackProviderContract
 
 /**
@@ -28,7 +28,7 @@ class LocationFragment(
     private lateinit var binding: DialogLocationBinding
 
 
-    private val mTimerObserver = object : VerifySessionTimer.Timer.TimerObserver {
+    private val mTimerObserver = object : DocumentSubmissionTimer.Observer {
         override fun onTick(millisRemaining: Int) {
             binding?.txtTimeRemaining?.text = getTimeRemainingString(millisRemaining, R.string.idv_timer_label)
         }
@@ -37,7 +37,7 @@ class LocationFragment(
 
     override fun onDestroyView() {
         super.onDestroyView()
-        VerifySessionTimer.getInstance().getSessionTimer().removeObserver(mTimerObserver)
+        DocumentSubmissionTimer.getInstance().removeObserver(mTimerObserver)
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
@@ -49,7 +49,7 @@ class LocationFragment(
             btnCancel.setOnClickListener(this@LocationFragment)
             btnSkip.setOnClickListener(this@LocationFragment)
             btnSkip.visibility = if (captureSettings.isOptional) View.VISIBLE else View.GONE
-            VerifySessionTimer.getInstance().getSessionTimer().addObserver(mTimerObserver)
+            DocumentSubmissionTimer.getInstance().addObserver(mTimerObserver)
         }
         return binding.root
     }